--- src/assets/tag/letter/send.tsx
@@ -1,25 +1,0 @@
-function SvgLetterStateSend() {
-  return (
-    <svg
-      width="4.2rem"
-      viewBox="0 0 58 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      role="img"
-      aria-labelledby="태그 - 편지 도착!"
-    >
-      <title>태그 - 편지 도착!</title>
-      <path
-        d="M0 10C0 4.477 4.477 0 10 0h38c5.523 0 10 4.477 10 10s-4.477 10-10 10H10C4.477 20 0 15.523 0 10z"
-        fill="#DCAE5C"
-      />
-      <path
-        d="M8.27 9.89V8.51h.74l.17-1.81h1.51l-.17 1.81h.79l.18-1.81h1.49l-.17 1.81h.71v-.73h.94V7.3h-.94v-.9H8.27V5.02h5.42v.88h.77v-.89h1.62v5.85h-1.62V9.18h-.77v.71H8.27zm7.95 2.32v1.38H9.08v-3.05h1.62v1.67h5.52zm4.844-5.8l1.44 6.86h-1.5l-.83-4.06-.37-1.99-.32 1.99-.83 4.06h-1.69l1.51-6.86h-1.38V5.03h5.29v1.38h-1.32zm3.87 6.96h-1.62V5.01h1.62v8.36zm6.76-1.38v-1.64h-2.59V5.02h6.83V6.4h-5.21v2.56h5.3v1.39h-2.71v1.64h3.23v1.38h-8.05v-1.38h3.2zm10.354-4.96h-.83l.84 2.68h-1.47l-.67-2.04-.17-.59-.19.56-.94 2.1h-1.79l1.2-2.71h-1.09V5.71h1.71l.17-.83h2.09l-.15.83h1.29v1.32zm.8 4.61h-5.48v-1.38h7.1v3.33h-1.62v-1.95zm2.72-3.7h-1.13V9.8h-1.62V5.01h1.62v1.55h1.13v1.38zm2.684-3.11l-.24 6.25h-1.29l-.22-6.25h1.75zm-1.76 7.02h1.67v1.73h-1.67v-1.73z"
-        fill="#11373E"
-      />
-    </svg>
-  );
-}
-
-export default SvgLetterStateSend;
-
